+++ src/Application/Adapter/NotificationSubscriberInterface.php
@@ -3,6 +3,7 @@
 namespace Proximum\Vimeet\Application\Adapter;
 
 
+use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\Happening;
 use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\User;
@@ -17,7 +18,9 @@ interface NotificationSubscriberInterface
 
     public function getNetworkingTopic(int $eventId): string;
 
-    public function getUserTopic(int $userId): string;
+    public function getUserTopic(int $eventId, int $userId): string;
+
+    public function getCallVisioTopic(Event $event): string;
 
     public function getUserSubscriberKey(Sheet $sheet, User $user): string;
 }